requirements.txt
fastapi~=0.115.3
uvicorn~=0.32.0
asyncpg
pydantic~=2.10.5
python-multipart~=0.0.19
python-dotenv~=1.0.1
contextvars
aiosmtplib
pyJWT
boto3
